+++ components/ledger/ledger-consensual-flow/build.gradle
@@ -16,6 +16,7 @@ dependencies {
     implementation project(":libs:crypto:crypto-core")
     implementation project(":libs:flows:flow-api")
     implementation project(":components:ledger:ledger-common-flow-api")
+    implementation project(":components:ledger:ledger-common-flow")
     implementation project(":components:flow:flow-service")
     implementation project(":libs:serialization:serialization-checkpoint-api")
     implementation project(":libs:virtual-node:sandbox-group-context")
