+++ components/ledger/ledger-utxo-flow/build.gradle
@@ -14,6 +14,7 @@ dependencies {
     implementation project(":libs:crypto:crypto-core")
     implementation project(":libs:ledger:ledger-common-data")
     implementation project(":libs:ledger:ledger-utxo-data")
+    implementation project(":components:ledger:ledger-common-flow")
     implementation project(":libs:sandbox")
 
     testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
