+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/impl/transaction/UtxoTransactionBuilderImpl.kt
@@ -4,7 +4,7 @@ import net.corda.ledger.common.data.transaction.CordaPackageSummary
 import net.corda.ledger.common.data.transaction.PrivacySaltImpl
 import net.corda.ledger.common.data.transaction.TransactionMetaData
 import net.corda.ledger.common.data.transaction.WireTransaction
-import net.corda.ledger.common.data.transaction.createTransactionSignature
+import net.corda.ledger.common.flow.impl.transaction.createTransactionSignature
 import net.corda.ledger.utxo.data.state.TransactionStateImpl
 import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup
 import net.corda.ledger.utxo.data.transaction.UtxoOutputInfoComponent
